Refactor
 - New command structur for sparating of informations
 - Add new control cmd
