+++ src/flee_and_catch/backend/communication/command/Connection.java
@@ -3,10 +3,10 @@ package flee_and_catch.backend.communication.command;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import flee_and_catch.backend.communication.identification.ClientIdentification;
-import flee_and_catch.backend.device.Device;
-import flee_and_catch.backend.device.app.App;
-import flee_and_catch.backend.device.robot.Robot;
+import flee_and_catch.backend.communication.command.device.Device;
+import flee_and_catch.backend.communication.command.device.app.App;
+import flee_and_catch.backend.communication.command.device.robot.Robot;
+import flee_and_catch.backend.communication.command.identification.ClientIdentification;
 
 public class Connection extends Command {	
 	private Device device;
