+++ src/flee_and_catch/backend/communication/command/ControlType.java
@@ -1,5 +1,5 @@
 package flee_and_catch.backend.communication.command;
 
 public enum ControlType {
-	Begin, End, Start, Stop
+	Begin, End, Start, Stop, Control
 }
