+++ src/flee_and_catch/backend/communication/command/component/IdentificationType.java
@@ -1,0 +1,5 @@
+package flee_and_catch.backend.communication.command.component;
+
+public enum IdentificationType {
+	App, Robot;
+}
