+++ src/flee_and_catch/backend/communication/command/component/RobotType.java
@@ -1,4 +1,4 @@
-package flee_and_catch.backend.component;
+package flee_and_catch.backend.communication.command.component;
 
 public enum RobotType {
 	ThreeWheelDrive, FourWheelDrive, ChainDrive;
