+++ src/flee_and_catch/backend/communication/command/device/Device.java
@@ -1,0 +1,4 @@
+package flee_and_catch.backend.communication.command.device;
+
+public interface Device {
+}
