+++ src/flee_and_catch/backend/communication/command/identification/ClientIdentification.java
@@ -1,4 +1,4 @@
-package flee_and_catch.backend.communication.identification;
+package flee_and_catch.backend.communication.command.identification;
 
 import org.json.JSONException;
 import org.json.JSONObject;
