--- src/flee_and_catch/backend/component/AppType.java
@@ -1,5 +1,0 @@
-package flee_and_catch.backend.component;
-
-public enum AppType {
-	App;
-}
