--- src/flee_and_catch/backend/component/IdentificationType.java
@@ -1,5 +1,0 @@
-package flee_and_catch.backend.component;
-
-public enum IdentificationType {
-	App, Robot;
-}
