--- src/flee_and_catch/backend/device/Device.java
@@ -1,4 +1,0 @@
-package flee_and_catch.backend.device;
-
-public interface Device {
-}
